refactor: finalize Spotify service unification
Finalized the Spotify service refactoring and addressed PR feedback.
- Verified type safety and error handling.
- Confirmed consolidation of logic into `SpotifyService`.
- Ensured consistent `deviceId` usage.
