--- services/spotifyUtils.ts
@@ -1,6 +1,0 @@
-export function isEmptyResponseError(error: unknown): boolean {
-  if (!(error instanceof SyntaxError)) {
-    return false
-  }
-  return /unexpected end of/i.test(error.message)
-}
