--- tests/unit/spotifyPlayerManager.test.ts
@@ -1,362 +1,0 @@
-import { describe, expect, it, jest, beforeEach } from '@jest/globals'
-import { SpotifyPlayerManager } from '../../services/spotifyPlayerManager'
-import { ServerMessage, SpotifyData } from '../../types/websocket'
-import { mockPlayer } from './spotify-test-utils'
-import { SpotifyApi } from '@spotify/web-api-ts-sdk'
-
-const NOT_PLAYING_MESSAGE = 'Nothing is currently playing.'
-
-// Mock logger
-jest.mock('../../utils/logger.server.js', () => ({
-  __esModule: true,
-  // eslint-disable-next-line @typescript-eslint/no-require-imports
-  default: require('./spotify-mocks').mockLogger,
-}))
-
-describe('SpotifyPlayerManager', () => {
-  let playerManager: SpotifyPlayerManager
-  let broadcastMock: jest.Mock<(message: ServerMessage) => void>
-  let getStateMock: jest.Mock<() => SpotifyData>
-  let setStateMock: jest.Mock<
-    (update: SpotifyData | ((prevState: SpotifyData) => SpotifyData)) => void
-  >
-  let currentState: SpotifyData
-
-  beforeEach(() => {
-    jest.clearAllMocks()
-
-    // Default state for testing
-    currentState = {
-      devices: [],
-      playback: {
-        track: {
-          id: null,
-          name: NOT_PLAYING_MESSAGE,
-          artist: '',
-          albumName: '',
-          albumArtUrl: '',
-        },
-        is_playing: false,
-        isMuted: false,
-        volume_percent: 70,
-        progress_ms: 0,
-      },
-    }
-
-    broadcastMock = jest.fn()
-    getStateMock = jest.fn(() => currentState)
-    setStateMock = jest.fn((update) => {
-      if (typeof update === 'function') {
-        currentState = update(currentState)
-      } else {
-        currentState = update
-      }
-    })
-
-    // Create a mock SDK using the mockPlayer from utilities
-    const mockSdk = {
-      player: mockPlayer,
-    } as unknown as SpotifyApi
-
-    playerManager = new SpotifyPlayerManager(
-      mockSdk,
-      broadcastMock,
-      getStateMock,
-      setStateMock
-    )
-  })
-
-  describe('refreshPlaybackState', () => {
-    it('should update state and broadcast when track changes', async () => {
-      // Arrange: SDK returns a new track
-      mockPlayer.getPlaybackState.mockResolvedValue({
-        device: { volume_percent: 70 },
-        is_playing: true,
-        progress_ms: 1000,
-        item: {
-          id: 'track1',
-          name: 'New Song',
-          type: 'track',
-          artists: [{ name: 'Artist 1' }],
-          album: { name: 'Album 1', images: [{ url: 'url1' }] },
-        },
-      })
-
-      // Act
-      await playerManager.refreshPlaybackState()
-
-      // Assert
-      // Verify state was updated
-      expect(setStateMock).toHaveBeenCalled()
-      // Verify broadcast was sent with correct details
-      expect(broadcastMock).toHaveBeenCalledWith({
-        type: 'SPOTIFY_UPDATE',
-        payload: expect.objectContaining({
-          playback: expect.objectContaining({
-            track: expect.objectContaining({
-              id: 'track1',
-              name: 'New Song',
-            }),
-            is_playing: true,
-          }),
-        }),
-      })
-    })
-
-    it('should update state and broadcast when playback stops', async () => {
-      // Arrange: Current state shows playing, SDK returns null
-      currentState = {
-        ...currentState,
-        playback: {
-          ...currentState.playback,
-          track: {
-            ...currentState.playback.track,
-            id: 'track1',
-            name: 'Some Song',
-          },
-          is_playing: true,
-        },
-      }
-      mockPlayer.getPlaybackState.mockResolvedValue(null)
-
-      // Act
-      await playerManager.refreshPlaybackState()
-
-      // Assert
-      expect(setStateMock).toHaveBeenCalled()
-      expect(broadcastMock).toHaveBeenCalledWith({
-        type: 'SPOTIFY_UPDATE',
-        payload: expect.objectContaining({
-          playback: expect.objectContaining({
-            track: expect.objectContaining({
-              id: null,
-              name: NOT_PLAYING_MESSAGE,
-            }),
-            is_playing: false,
-          }),
-        }),
-      })
-    })
-
-    it('should NOT update or broadcast if state is unchanged (playing same song)', async () => {
-      // Arrange: Current state matches SDK response
-      currentState = {
-        ...currentState,
-        playback: {
-          track: {
-            id: 'track1',
-            name: 'Song 1',
-            artist: 'Artist 1',
-            albumName: 'Album 1',
-            albumArtUrl: 'url1',
-          },
-          is_playing: true,
-          volume_percent: 70,
-          progress_ms: 1000,
-        },
-      }
-
-      mockPlayer.getPlaybackState.mockResolvedValue({
-        device: { volume_percent: 70 },
-        is_playing: true,
-        progress_ms: 1000,
-        item: {
-          id: 'track1',
-          name: 'Song 1',
-          type: 'track',
-          artists: [{ name: 'Artist 1' }],
-          album: { name: 'Album 1', images: [{ url: 'url1' }] },
-        },
-      })
-
-      // Act
-      await playerManager.refreshPlaybackState()
-
-      // Assert
-      expect(setStateMock).not.toHaveBeenCalled()
-      expect(broadcastMock).not.toHaveBeenCalled()
-    })
-
-    it('should NOT update or broadcast if state is unchanged (stopped)', async () => {
-      // Arrange: Current state is stopped, SDK returns null
-      currentState = {
-        ...currentState,
-        playback: {
-          ...currentState.playback,
-          track: {
-            ...currentState.playback.track,
-            id: null,
-            name: NOT_PLAYING_MESSAGE,
-          },
-          is_playing: false,
-        },
-      }
-      mockPlayer.getPlaybackState.mockResolvedValue(null)
-
-      // Act
-      await playerManager.refreshPlaybackState()
-
-      // Assert
-      expect(setStateMock).not.toHaveBeenCalled()
-      expect(broadcastMock).not.toHaveBeenCalled()
-    })
-
-    it('should handle podcast episodes correctly', async () => {
-      // Arrange: SDK returns an episode
-      mockPlayer.getPlaybackState.mockResolvedValue({
-        device: { volume_percent: 70 },
-        is_playing: true,
-        progress_ms: 1000,
-        item: {
-          id: 'episode1',
-          name: 'Podcast Episode',
-          type: 'episode',
-          show: {
-            publisher: 'Podcast Host',
-            name: 'Podcast Show',
-            images: [{ url: 'podcast_url' }],
-          },
-        },
-      })
-
-      // Act
-      await playerManager.refreshPlaybackState()
-
-      // Assert
-      expect(broadcastMock).toHaveBeenCalledWith({
-        type: 'SPOTIFY_UPDATE',
-        payload: expect.objectContaining({
-          playback: expect.objectContaining({
-            track: expect.objectContaining({
-              id: 'episode1',
-              name: 'Podcast Episode',
-            }),
-          }),
-        }),
-      })
-    })
-
-    it('should handle null/missing item safely', async () => {
-      // Arrange: SDK returns response with null item
-      mockPlayer.getPlaybackState.mockResolvedValue({
-        device: { volume_percent: 70 },
-        item: null,
-        is_playing: false,
-      })
-
-      currentState = {
-        ...currentState,
-        playback: {
-          ...currentState.playback,
-          is_playing: true,
-          track: {
-            ...currentState.playback.track,
-            name: 'Previous Song',
-          },
-        },
-      }
-
-      // Act
-      await playerManager.refreshPlaybackState()
-
-      // Assert
-      expect(setStateMock).toHaveBeenCalled()
-      expect(broadcastMock).toHaveBeenCalledWith({
-        type: 'SPOTIFY_UPDATE',
-        payload: expect.objectContaining({
-          playback: expect.objectContaining({
-            track: expect.objectContaining({
-              id: null,
-              name: NOT_PLAYING_MESSAGE,
-            }),
-            is_playing: false,
-          }),
-        }),
-      })
-    })
-  })
-
-  describe('executeSpotifyCommand', () => {
-    it('should handle PLAY command', async () => {
-      await playerManager.executeSpotifyCommand('PLAY', {
-        deviceId: 'device_id',
-      })
-      expect(mockPlayer.startResumePlayback).toHaveBeenCalledWith('device_id')
-    })
-
-    it('should handle PAUSE command', async () => {
-      await playerManager.executeSpotifyCommand('PAUSE', {
-        deviceId: 'device_id',
-      })
-      expect(mockPlayer.pausePlayback).toHaveBeenCalledWith('device_id')
-    })
-
-    it('should handle NEXT command', async () => {
-      await playerManager.executeSpotifyCommand('NEXT', {
-        deviceId: 'device_id',
-      })
-      expect(mockPlayer.skipToNext).toHaveBeenCalledWith('device_id')
-    })
-
-    it('should handle PREVIOUS command', async () => {
-      await playerManager.executeSpotifyCommand('PREVIOUS', {
-        deviceId: 'device_id',
-      })
-      expect(mockPlayer.skipToPrevious).toHaveBeenCalledWith('device_id')
-    })
-
-    it('should handle SET_VOLUME and update local state', async () => {
-      await playerManager.executeSpotifyCommand('SET_VOLUME', {
-        deviceId: 'device_id',
-        volume: 50,
-      })
-      expect(mockPlayer.setPlaybackVolume).toHaveBeenCalledWith(50, 'device_id')
-      expect(setStateMock).toHaveBeenCalled()
-      expect(currentState.playback.volume_percent).toBe(50)
-      expect(currentState.playback.isMuted).toBe(false)
-    })
-
-    it('should clamp volume correctly', async () => {
-      await playerManager.executeSpotifyCommand('SET_VOLUME', {
-        deviceId: 'device_id',
-        volume: 150,
-      })
-      expect(mockPlayer.setPlaybackVolume).toHaveBeenCalledWith(
-        100,
-        'device_id'
-      )
-      expect(currentState.playback.volume_percent).toBe(100)
-
-      await playerManager.executeSpotifyCommand('SET_VOLUME', {
-        deviceId: 'device_id',
-        volume: -10,
-      })
-      expect(mockPlayer.setPlaybackVolume).toHaveBeenCalledWith(0, 'device_id')
-      expect(currentState.playback.volume_percent).toBe(0)
-      expect(currentState.playback.isMuted).toBe(true)
-    })
-
-    it('should handle TRANSFER_PLAYBACK', async () => {
-      await playerManager.executeSpotifyCommand('TRANSFER_PLAYBACK', {
-        deviceId: 'new_device_id',
-      })
-      expect(mockPlayer.transferPlayback).toHaveBeenCalledWith(
-        ['new_device_id'],
-        true
-      )
-    })
-
-    it('should log debug on 204 No Content', async () => {
-      mockPlayer.startResumePlayback.mockRejectedValue(new Error('')) // Simulate empty error often seen with 204
-      // We can't easily mock the internal isEmptyResponseError utility to true without more setup,
-      // but we can ensure it handles errors gracefully.
-      // For this test, we'll simulate a standard API error to ensure error logging works,
-      // as 204 handling is strictly an integration detail of the SDK wrapper/utility.
-      mockPlayer.startResumePlayback.mockRejectedValue(new Error('API Error'))
-
-      await expect(
-        playerManager.executeSpotifyCommand('PLAY', { deviceId: 'd1' })
-      ).rejects.toThrow('API Error')
-    })
-  })
-})
